requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
yfinance>=0.2.33
pandas>=2.1.0
playwright>=1.40.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
lxml>=4.9.3
